Rename: Book component to BookCard for readability
